regex validations

+++ app.py
@@ -4,6 +4,7 @@ from dotenv import load_dotenv
 import os
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_wtf.csrf import CSRFProtect
+import re
 
 load_dotenv()
 
@@ -21,6 +22,7 @@ mydb = mysql.connector.connect(
     auth_plugin="mysql_native_password"
 )
 
+ALLOWED_PATTERN = re.compile(r"^[a-zA-Z0-9\s.,!?'-]{1,200}$")
 
 @app.route('/')
 def signup():
@@ -36,6 +38,14 @@ def validate_info():
     check_account_query = "SELECT * FROM login WHERE email=%s OR username=%s"
     mycursor.execute(check_account_query, (email, username))
     user = mycursor.fetchone()
+
+    if not ALLOWED_PATTERN.match(username):
+        flash("Invalid Username description. Please avoid using special characters.")
+        return redirect('/')
+
+    if len(password) < 8 or len(password) > 12:
+        flash("Password must be between 8 and 12 characters long.")
+        return redirect('/')
 
     if user:
         flash("This Account Already Exist!")
@@ -65,6 +75,10 @@ def validate_login():
     mycursor.execute(verify_account_query, (username,))
     user = mycursor.fetchone()
     mycursor.close()
+
+    if not ALLOWED_PATTERN.match(username):
+        flash("Invalid credentials description. Please avoid using special characters.")
+        return redirect('/l')
 
     if user and check_password_hash(user[2], password):
         session['username'] = username
@@ -99,74 +113,96 @@ def tasks():
     
         return render_template("tasks.html", user = username, task = tasks, fav = favs)
     else:
+        flask("First Login To Access Features.")
         return redirect('/l')
 
 @app.route('/add', methods = ['POST'])
 def add():
-    username = session['username']
-    task = request.form['task']
-    mycursor = mydb.cursor()
+    if "username" in session:
+        username = session['username']
+        task = request.form['task']
+        mycursor = mydb.cursor()
 
-    select_tasks_query = "SELECT description, status, fav FROM tasks WHERE username=%s AND description=%s"
-    mycursor.execute(select_tasks_query, (username, task))
-    exist = mycursor.fetchone()
-    
-    if exist:
-        flash("This task is already present")
+        select_tasks_query = "SELECT description, status, fav FROM tasks WHERE username=%s AND description=%s"
+        mycursor.execute(select_tasks_query, (username, task))
+        exist = mycursor.fetchone()
+
+        if not ALLOWED_PATTERN.match(task):
+            flash("Invalid task description. Please avoid using special characters.")
+            return redirect('/task')
         
+        if exist:
+            flash("This task is already present")
+            
+        else:
+            add_task_query = "INSERT INTO tasks (username, description) VALUES(%s, %s)"
+            mycursor.execute(add_task_query, (username, task))
+            mydb.commit()
+        
+        mycursor.close()
+        return redirect('/task')
     else:
-        add_task_query = "INSERT INTO tasks (username, description) VALUES(%s, %s)"
-        mycursor.execute(add_task_query, (username, task))
-        mydb.commit()
-    
-    mycursor.close()
-    return redirect('/task')
+        flask("First Login To Access Features.", "info")
+        return redirect('/l')
+
 
 @app.route('/delete', methods = ['POST'])
 def delete():
-    username = session['username']
-    t1 = request.form['t1']
-    mycursor = mydb.cursor()
+    if "username" in session:
+        username = session['username']
+        t1 = request.form['t1']
+        mycursor = mydb.cursor()
 
-    delete_task_query = "DELETE FROM tasks WHERE username=%s AND description=%s"
-    mycursor.execute(delete_task_query, (username, t1))
-    mydb.commit()
-    mycursor.close()
+        delete_task_query = "DELETE FROM tasks WHERE username=%s AND description=%s"
+        mycursor.execute(delete_task_query, (username, t1))
+        mydb.commit()
+        mycursor.close()
 
-    return redirect('/task')
+        return redirect('/task')
+    else:
+        flask("First Login To Access Features.", "info")
+        return redirect('/l')
 
 @app.route('/update_status', methods=['POST'])
 def update_status():
-    username = session['username']
-    task = request.form['t1']
-    status = request.form['status']
-    mycursor = mydb.cursor()
+    if "username" in session:
+        username = session['username']
+        task = request.form['t1']
+        status = request.form['status']
+        mycursor = mydb.cursor()
 
-    update_status_query = "UPDATE tasks SET status = %s WHERE username = %s AND description = %s"
-    mycursor.execute(update_status_query, (status, username, task))
-    mydb.commit()
-    mycursor.close()
+        update_status_query = "UPDATE tasks SET status = %s WHERE username = %s AND description = %s"
+        mycursor.execute(update_status_query, (status, username, task))
+        mydb.commit()
+        mycursor.close()
 
-    return redirect('/task')
+        return redirect('/task')
+    else:
+        flask("First Login To Access Features.", "info")
+        return redirect('/l')
 
 @app.route('/fav', methods = ['Post'])
 def daily_task():
-    username = session['username']
-    task = request.form['t1']
-    mycursor = mydb.cursor()
+    if "username" in session:
+        username = session['username']
+        task = request.form['t1']
+        mycursor = mydb.cursor()
 
-    add_dailytasks_query = "SELECT fav FROM tasks WHERE username=%s AND description=%s"
-    mycursor.execute(add_dailytasks_query, (username, task))
-    current = mycursor.fetchone()
+        add_dailytasks_query = "SELECT fav FROM tasks WHERE username=%s AND description=%s"
+        mycursor.execute(add_dailytasks_query, (username, task))
+        current = mycursor.fetchone()
 
-    if current:
-        new_fav = 0 if current[0] == 1 else 1
-        update_dailytasks_query = "UPDATE tasks SET fav=%s WHERE username=%s AND description=%s"
-        mycursor.execute(update_dailytasks_query, (new_fav, username, task))
-        mydb.commit()
+        if current:
+            new_fav = 0 if current[0] == 1 else 1
+            update_dailytasks_query = "UPDATE tasks SET fav=%s WHERE username=%s AND description=%s"
+            mycursor.execute(update_dailytasks_query, (new_fav, username, task))
+            mydb.commit()
 
-    mycursor.close()
-    return redirect('/task')
+        mycursor.close()
+        return redirect('/task')
+    else:
+        flask("First Login To Access Features.", "info")
+        return redirect('/l')
 
 @app.route('/show')
 def show_daily():
@@ -187,25 +223,32 @@ def show_daily():
         if results:
             flash("Here You Go, Your Daily Tasks.")
         return render_template("daily.html", task = tasks, fav = favs, user = username)
+    else:
+        flask("First Login To Access Features.")
+        return redirect('/l')
 
 @app.route('/remove', methods=['Post'])
 def remove_dailytask():
-    username = session['username']
-    task = request.form['t1']
-    mycursor = mydb.cursor()
-    
-    access_dailytasks_query = "SELECT fav FROM tasks WHERE username=%s AND description=%s"
-    mycursor.execute(access_dailytasks_query, (username, task))
-    current = mycursor.fetchone()
+    if "username" in session:
+        username = session['username']
+        task = request.form['t1']
+        mycursor = mydb.cursor()
 
-    if current:
-        new_fav = 0 
-        remove_dailytask_query = "UPDATE tasks SET fav = %s WHERE username = %s AND description = %s"
-        mycursor.execute(remove_dailytask_query, (new_fav, username, task))
-        mydb.commit()
-    
-    mycursor.close()
-    return redirect('/show')
+        access_dailytasks_query = "SELECT fav FROM tasks WHERE username=%s AND description=%s"
+        mycursor.execute(access_dailytasks_query, (username, task))
+        current = mycursor.fetchone()
+
+        if current:
+            new_fav = 0 
+            remove_dailytask_query = "UPDATE tasks SET fav = %s WHERE username = %s AND description = %s"
+            mycursor.execute(remove_dailytask_query, (new_fav, username, task))
+            mydb.commit()
+
+        mycursor.close()
+        return redirect('/show')
+    else:
+        flask("First Login To Access Features.")
+        return redirect('/l')
 
 @app.route('/logout')
 def logout():
@@ -214,22 +257,26 @@ def logout():
 
 @app.route('/delete_account')
 def delete_account():
-    username = session['username']
-    mycursor = mydb.cursor()
+    if "username" in session:
+        username = session['username']
+        mycursor = mydb.cursor()
 
-    delete_account_query = "DELETE FROM login WHERE username=%s"
-    mycursor.execute(delete_account_query, (username,))
-    mydb.commit()
+        delete_account_query = "DELETE FROM login WHERE username=%s"
+        mycursor.execute(delete_account_query, (username,))
+        mydb.commit()
 
-    delete_alltasks_query = "DELETE FROM tasks WHERE username=%s"
-    mycursor.execute(delete_alltasks_query, (username,))
-    mydb.commit()
-    
-    mycursor.close()
-    session.pop("username", None)
-    
-    flash("Account Was Deleted Successfully, SignUp To Create New", "info")
-    return redirect('/')
+        delete_alltasks_query = "DELETE FROM tasks WHERE username=%s"
+        mycursor.execute(delete_alltasks_query, (username,))
+        mydb.commit()
+        
+        mycursor.close()
+        session.pop("username", None)
+        
+        flash("Account Was Deleted Successfully, SignUp To Create New", "info")
+        return redirect('/')
+    else:
+        flask("First Login To Access Features.","info")
+        return redirect('/l')
 
 
 if __name__ == '__main__':
